+++ inform/__init__.py
@@ -1,5 +1,5 @@
-__version__ = '1.13.2'
-__released__ = '2018-09-09'
+__version__ = '1.13.3'
+__released__ = '2018-10-07'
 
 from .inform import (
     # inform utility functions